pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "python-action"
version = "0.1.0"
description = "GitHub Actions Python Template"
license = "MIT"
dependencies = [
    "actions-python-core",
]
[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "ruff",
]

[tool.ruff]
line-length = 88
select = ["B", "E", "F", "I", "N", "RUF"]
src = ["packages/*"]

[tool.ruff.isort]
combine-as-imports = true

[tool.coverage.report]
exclude_also = [
    "if __name__ == .__main__.:",
    "if (typing\\.)?TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.run]
branch = true
